# Improved performance of GenerateArrayStrings