Hello!

With these little changes, the performance of map generation would improve from horrendous to acceptable.

From my testing I found out that MapFileTools.GenerateArrayStrings() is responsible for like 99% of the total time it takes to generate a map. Some stats from my testing
200x200 map:
    Before: 21.26 seconds to create a map
    After: 1.49 seconds to create a map
384x384 map:
    Before: 4 minutes 28 seconds to create a map
    After: 6.02 seconds to create a map

These times do not take into account the loading of the map, just the actual generation.