@R-T-B 2022-08-27
This is excellent!  I had been led to believe the issue was at least mostly the use of dictionaries to serialize the JSON.  I'm sure that's still a potential problem, but this shows that massive gains were to be had in that method.  Great work!  I will merge this immediately, and pending a few tests and benches, release it promptly. :)